pyproject.toml
[tool.poetry]
name = "supercomposed"
version = "0.1.0"
description = ""
authors = ["Burak Yigit Kaya <byk@sentry.io>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
types-PyYAML = "^6.0.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
